// src/Blindating/scripts/models/preference.ts
export class Preference {
    id:           number;
    gender:       string;
    relationship: string;
    from:         string;
    to:           string;
    country:      string;
    city:         string;
    hcolor:       string;
    ecolor:       string;
    hobby:        string;
    vkontakte:    string;
    facebook:     string;
    twitter:      string;
    google:       string;
}